My Files So Far
just a bunch of code (programs) that i've done so far. the code i do from the lessons goes in the "Learning.py" file, the projects that codecademy gives me go in their own file. the "js-->py.py" file is for when i finish the python course on codecademy, i will start rewriting all the code i have in my Javascript folder over to python.