Added New Project (Sal's Shipping)
More python from codecademy's course, but it won't work in vscode for some reason. i could change it, but it would look bad and would be harder to read. I might not be able to fix it though actually. It works in codecademy, so i think that i have a version of python too advanced to even understand this primitive code.